requirements.txt
fzflib
rich
python-vlc
requests